+++ R/draw_legend_utils.R
@@ -57,38 +57,29 @@ compute_legend_args = function(
   legend_args[["bty"]] = legend_args[["bty"]] %||% "n"
   legend_args[["horiz"]] = legend_args[["horiz"]] %||% FALSE
   legend_args[["xpd"]] = legend_args[["xpd"]] %||% NA
-  if (!isTRUE(type %in% c("p", "ribbon", "polygon", "polypath"))) {
-    legend_args[["lwd"]] = legend_args[["lwd"]] %||% lwd
-  }
-  if (is.null(type) || type %in% c("p", "pointrange", "errorbar", "text")) {
+  legend_args[["lwd"]] = legend_args[["lwd"]] %||% lwd
+  # special handling of pt.cex for bubble plots
+  # (fixme: can't handle ahead of time in bubble.R b/c of dual legend gotcha)
+  if (is.null(type) || type %in% c("p", "text")) {
     legend_args[["pt.cex"]] = legend_args[["pt.cex"]] %||% (cex %||% par("cex"))
   }
-  # turn off inner line for "barplot" type
-  if (identical(type, "barplot")) {
-    legend_args[["lty"]] = 0
-  }
-  if (isTRUE(type %in% c("rect", "ribbon", "polygon", "polypath", "boxplot", "hist", "histogram", "spineplot", "ridge", "barplot", "violin")) || gradient) {
+  if (gradient) {
     legend_args[["pch"]] = 22
     legend_args[["pt.cex"]] = legend_args[["pt.cex"]] %||% 3.5
     legend_args[["y.intersp"]] = legend_args[["y.intersp"]] %||% 1.25
     legend_args[["seg.len"]] = legend_args[["seg.len"]] %||% 1.25
   }
-  if (isTRUE(type %in% c("ribbon", "hist", "histogram", "spineplot"))) {
-    legend_args[["pt.lwd"]] = legend_args[["pt.lwd"]] %||% 0
-  }
-  if (identical(type, "p")) {
-    legend_args[["pt.lwd"]] = legend_args[["pt.lwd"]] %||% lwd
-  }
   if (identical(type, "n") && isFALSE(gradient)) {
     legend_args[["pch"]] = legend_args[["pch"]] %||% par("pch")
   }
+  # Special pt.bg handling for types that need color-based fills
   if (identical(type, "spineplot")) {
     legend_args[["pt.bg"]] = legend_args[["pt.bg"]] %||% legend_args[["col"]]
-  }
-  if (identical(type, "ridge") && isFALSE(gradient)) {
+  } else if (identical(type, "ridge") && isFALSE(gradient)) {
     legend_args[["pt.bg"]] = legend_args[["pt.bg"]] %||% sapply(legend_args[["col"]], function(ccol) seq_palette(ccol, n = 2)[2])
+  } else {
+    legend_args[["pt.bg"]] = legend_args[["pt.bg"]] %||% bg
   }
-  legend_args[["pt.bg"]] = legend_args[["pt.bg"]] %||% bg
   legend_args[["legend"]] = legend_args[["legend"]] %||% lgnd_labs
   if (length(lgnd_labs) != length(eval(legend_args[["legend"]]))) {
     warning(