+++ R/type_barplot.R
@@ -199,6 +199,14 @@ data_barplot = function(width = 5/6, beside = FALSE, center = FALSE, FUN = NULL,
         xaxs = "r"
         xaxt = if (xaxt == "s") "l" else xaxt
         yaxs = "i"
+        
+        # legend customizations
+        settings$legend_args[["lty"]] = settings$legend_args[["lty"]] %||% 0
+        settings$legend_args[["pch"]] = settings$legend_args[["pch"]] %||% 22
+        settings$legend_args[["pt.cex"]] = settings$legend_args[["pt.cex"]] %||% 3.5
+        settings$legend_args[["y.intersp"]] = settings$legend_args[["y.intersp"]] %||% 1.25
+        settings$legend_args[["seg.len"]] = settings$legend_args[["seg.len"]] %||% 1.25
+        
         env2env(environment(), settings, c(
           "datapoints",
           "xlab",