+++ R/type_ridge.R
@@ -409,6 +409,13 @@ data_ridge = function(bw = "nrd0", adjust = 1, kernel = "gaussian", n = 512,
       col = col,
       alpha = alpha
     )
+    
+    # legend customizations
+    settings$legend_args[["pch"]] = settings$legend_args[["pch"]] %||% 22
+    settings$legend_args[["pt.cex"]] = settings$legend_args[["pt.cex"]] %||% 3.5
+    settings$legend_args[["y.intersp"]] = settings$legend_args[["y.intersp"]] %||% 1.25
+    settings$legend_args[["seg.len"]] = settings$legend_args[["seg.len"]] %||% 1.25
+    
     env2env(environment(), settings, c(
       "datapoints",
       "yaxt",